+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes for each version of this project will be documented in this file.
 
+## 9.1.1
+
+### General
+- `IgxHierarchicalGrid`
+    - `onGridInitialized` - New output has been exposed. Emitted after a grid is being initialized for the corresponding row island.
+
 ## 9.1.0
 
 ### General